Move test database setup to fixtures module
Put all test fixtures in the `tests.fixtures` module for
discoverability.
